requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
Django==5.0.6
django-crontab==0.7.1
django-seed==0.3.1
et-xmlfile==1.1.0
Faker==27.4.0
fonttools==4.54.1
html5lib==1.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyparsing==3.1.4
pyphen==0.16.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
toposort==1.10
typing_extensions==4.12.2
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.2.3
